+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 * Enhancement: Improved `tracker.js` by switching to xhr/POST requests and adding helpful console logs.
 * Enhancement: Enhanced the Add-Ons page and license activation status display.
 * Enhancement: Adjusted the notices so that all are only shown to users with the `manage_options` capability, with some notices appearing only on the plugin pages.
+* Enhancement: Improved the posts report view, customized date format, and datepicker functionality.
+* Enhancement: Adjusted custom date range text to reflect WordPress settings.
+* Enhancement: Prevented Chart.js from being enqueued in the admin bar when the mini-chart add-on is disabled.
+* Enhancement: Enhanced table sorting functionality.
 * Enhancement: Implemented minor improvements.
 
 = 14.9.2 - 22.07.2024 =